2025-06-23_taller_visualizacion_datos_tiempo_real_graficas/requirements.txt
matplotlib
numpy
plotly
pandas
opencv-python
pillow
